Make video upload to S3 compatible with SigV4 authentication
The studio has a feature that allows course teams to upload videos
directly to S3. This is a very convenient feature, but it works only for
some S3 regions. The reason is that some regions require SigV4
authentication. It is possible to enable SigV4 authentication in boto by
setting the S3_USE_SIGV4 environment variable. However, this requires
the "host" setting to be passed to the boto.connection.Connection
constructor.

Actually, there are many other settings that should be passed to the
boto constructor, and this is precisely what the
storages.backends.s3.S3BotoStorage constructor does. So we replace the
call to boto by a call to the storages app.

All other calls to boto are made via the storages app in edx-platform.
As far as we know, the only exception left is in ORA2.

edX.org has not faced this problem because they (probably) use a
different authentication capability.

This change is important because it allows video uploads to, for
instance, the us-east-2 region.  The S3 emulator called MinIO (minio.io)
also requires SigV4. This emulator is useful for bringing object storage
capabilities to Open edX administrators which may not have access to
AWS.