refactor: Stop using globals_dict as output parameter in safe_exec
This mutation-based approach to returning a value is prone to developer
error. Additionally, it mildly complicates possible development of a
codejail that uses remote execution, which would be serializing the
globals for transport.

In fact, codejail's own `safe_exec` and `not_safe_exec` *already*
serialize the globals to JSON before passing them for execution, and
mutation of the original object only occurs at the very end in lieu of
a return statement.